LPS-21697 Test formatting
