+++ build-test.xml
@@ -275,7 +275,7 @@
 	</target>
 
 	<target name="build-test-ant-scripts">
-		<property name="build-test-ant-scripts.version" value="7" />
+		<property name="build-test-ant-scripts.version" value="6" />
 
 		<if>
 			<available file="portal-web/test-ant-scripts/version" />
@@ -1775,11 +1775,6 @@
 		<antcall target="build-test-ant-script">
 			<param name="test.ant.script.command" value="staging-community-sites" />
 			<param name="test.ant.script.suite" value="StagingCommunitySitesTestSuite" />
-		</antcall>
-
-		<antcall target="build-test-ant-script">
-			<param name="test.ant.script.command" value="staging-community-user" />
-			<param name="test.ant.script.suite" value="StagingCommunityUserTestSuite" />
 		</antcall>
 
 		<antcall target="build-test-ant-script">
