+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/ActivateStagingTest.html
@@ -185,7 +185,7 @@
 </tr>
 <tr>
 	<td>store</td>
-	<td>eval(selenium.isChecked('_165_staged-portlet_162Checkbox'))</td>
+	<td>eval(selenium.isChecked('_165_staged-portlet_19Checkbox'))</td>
 	<td>messageBoardsChecked</td>
 </tr>
 <tr>
@@ -195,7 +195,7 @@
 </tr>
 <tr>
 	<td>clickAt</td>
-	<td>//input[@id='_165_staged-portlet_162Checkbox']</td>
+	<td>//input[@id='_165_staged-portlet_19Checkbox']</td>
 	<td>Message Boards</td>
 </tr>
 <tr>
@@ -353,7 +353,6 @@
 	<td>//h1[@class='header-title']/span</td>
 	<td>Site Name (Staging)</td>
 </tr>
-
 </tbody></table>
 </body>
 </html>
