+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/AddPageSiteTest.html
@@ -39,6 +39,11 @@
 	<td>Search</td>
 </tr>
 <tr>
+	<td>assertText</td>
+	<td>//span[@title='Actions']/ul/li/strong/a</td>
+	<td>Actions</td>
+</tr>
+<tr>
 	<td>clickAt</td>
 	<td>//span[@title='Actions']/ul/li/strong/a</td>
 	<td>Actions</td>
@@ -83,7 +88,6 @@
 	<td>//div[@class='portlet-msg-success']</td>
 	<td>Your request completed successfully.</td>
 </tr>
-
 </tbody></table>
 </body>
 </html>
