+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/AddSiteTest.html
@@ -30,28 +30,28 @@
 </tr>
 <tr>
 	<td>assertText</td>
-	<td>//span[2]/ul/li/strong/a/span</td>
+	<td>//span[@title='Add']/ul/li/strong/a/span</td>
 	<td>Add</td>
 </tr>
 <tr>
 	<td>clickAt</td>
-	<td>//span[2]/ul/li/strong/a/span</td>
+	<td>//span[@title='Add']/ul/li/strong/a/span</td>
 	<td>Add</td>
 </tr>
 <tr>
 	<td>waitForVisible</td>
 	<td>//div[@class='lfr-component lfr-menu-list']/ul/li[1]/a</td>
-	<td>Site</td>
+	<td>Blank Site</td>
 </tr>
 <tr>
 	<td>assertText</td>
 	<td>//div[@class='lfr-component lfr-menu-list']/ul/li[1]/a</td>
-	<td>Site</td>
+	<td>Blank Site</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
 	<td>//div[@class='lfr-component lfr-menu-list']/ul/li[1]/a</td>
-	<td>Site</td>
+	<td>Blank Site</td>
 </tr>
 <tr>
 	<td>type</td>
@@ -78,7 +78,6 @@
 	<td>//h1[@class='header-title']/span</td>
 	<td>Site Name</td>
 </tr>
-
 </tbody></table>
 </body>
 </html>
