+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/DefineMemberRoleTest.html
