+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/User_AccessStagingTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Member_AccessStagingTest</title>
+<title>User_AccessStagingTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">Member_AccessStagingTest</td></tr>
+<tr><td rowspan="1" colspan="3">User_AccessStagingTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
