+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/User_AccessStagingTests.html
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
+<head>
+  <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
+  <title>User_AccessStagingTests</title>
+</head>
+<body>
+<table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
+<tr><td><b>User_AccessStagingTests</b></td></tr>
+<tr><td><a href="AddSiteTest.html">AddSiteTest</a></td></tr>
+<tr><td><a href="ActivateStagingTest.html">ActivateStagingTest</a></td></tr>
+<tr><td><a href="AddPageSiteTest.html">AddPageSiteTest</a></td></tr>
+<tr><td><a href="AddMemberTest.html">AddMemberTest</a></td></tr>
+<tr><td><a href="AddMemberRoleTest.html">AddMemberRoleTest</a></td></tr>
+<tr><td><a href="DefineMemberRoleTest.html">DefineMemberRoleTest</a></td></tr>
+<tr><td><a href="RemoveRolePowerUserTest.html">RemoveRolePowerUserTest</a></td></tr>
+<tr><td><a href="AssignUserRolesTest.html">AssignUserRolesTest</a></td></tr>
+<tr><td><a href="AssignMembersTest.html">AssignMembersTest</a></td></tr>
+<tr><td><a href="LogOutTest.html">LogOutTest</a></td></tr>
+<tr><td><a href="LoginUsersTest.html">LoginUsersTest</a></td></tr>
+<tr><td><a href="User_LoginTest.html">User_LoginTest</a></td></tr>
+<tr><td><a href="User_AccessStagingTest.html">User_AccessStagingTest</a></td></tr>
+<tr><td><a href="LogOutTest.html">LogOutTest</a></td></tr>
+<tr><td><a href="SA_LoginTest.html">SA_LoginTest</a></td></tr>
+<tr><td><a href="TearDownUserTest.html">TearDownUserTest</a></td></tr>
+<tr><td><a href="TearDownRolesTest.html">TearDownRolesTest</a></td></tr>
+<tr><td><a href="TearDownSitesTest.html">TearDownSitesTest</a></td></tr>
+</tbody></table>
+</body>
+</html>
