+++ portal-web/test/com/liferay/portalweb/stagingcommunity/sites/site/useraccessstaging/User_LoginTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Member_LoginTest</title>
+<title>User_LoginTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">Member_LoginTest</td></tr>
+<tr><td rowspan="1" colspan="3">User_LoginTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
